refactor ffi module
